Move parallel_execute to a new module.
